+++ android/app/src/main/java/org/westernghats/fieldvalidator/MainActivity.java
@@ -1,5 +1,13 @@
 package org.westernghats.fieldvalidator;
 
+import android.os.Bundle;
 import com.getcapacitor.BridgeActivity;
 
-public class MainActivity extends BridgeActivity {}
+public class MainActivity extends BridgeActivity {
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        registerPlugin(TrackStatusPlugin.class);
+        super.onCreate(savedInstanceState);
+    }
+}